Move ref forwarding tests to ref-forwarded.test.tsx
